use initial.json instead of initial_data.json
There was a reason why initial_data.json was renamed to initial.json:
initial_data is autoloaded every time migrations are run, which was causing
some data to be lost: in this case, it would switch to the default template,
when some deploys for clients had another template.
(Django is deprecating that automagic soon anyway, but we didn't want it here.)

However, we do still want to load that data in a new instance, hence
this commit. Just renaming initial_data to initial left new dev instances in
a dumb state (e.g.: ValueError at /account/login/ - The 'logo_image' attribute
has no file associated with it.)

Unfortunately, ./manage.py loaddata some_invalid_fixture_name actually
runs with exit code 0, so the scripts weren't halting as they should have
in order to highlight the problem, but what can you do