avoid empty dd tags in metadata detail templates
When values like email and location were empty strings (possibly because they
have string type in the database), the template was rendering <dd></dd>,
which meant that all the other <dd> elements were pushing up the page,
along with the following span.subtitle.

Why fix this at the Django template level?

1. It doesn't seem useful to ensure the database has NULLs or that we intervene
at the Python level (before template rendering) to translate null strings into
None.

2. We probably either want to delete both dt and dd, or to explicitly indicate
to user that there is no value. So if in the future there is a solution which
doesn't require adding presentation-only classes to dd tags, that could be
worth taking, but we still need the template-level fix.