django-osgeo-importer: add settings
These settings are needed in Exchange as a prerequisite to enabling
django-osgeo-importer to be used with Exchange.

OSGEO_DATASTORE tells django-osgeo-importer to use the exchange database and
user, in lieu of creating all new stuff in postgres. We can only avoid this if
we create a different database to match django-osgeo-importer defaults.

OSGEO_IMPORTER_GEONODE_ENABLED is a setting defined by django-osgeo-importer to
allow compatibility with GeoNode, notably to use the Profile model instead of
User. See the README.md for django-osgeo-importer for documentation on that. We
have to set this in Exchange unless it becomes the upstream default.

CELERY_IMPORTS also has to be adjusted to ensure that celery imports the
django-osgeo-importer celery tasks that are required for it to function.
CELERY_IMPORTS itself is explicitly imported to make this file easier to read
by humans and static checkers (even if it does come in via import *).

CELERY_IGNORE_RESULT is used by django-osgeo-importer for the handler
which reports to the UI when background processing is done. Without this
change, the GeoNode default value causes the UI polling to repeat forever.