adding  a specific fixture for oauth2_providor - specific for docker
Added new additions for login and logout endpoints to `OGC_SERVER` dictionary.  Added required `AUTH_EXEMPT_URLS` that are required for geoserver oauth extension to communicate to geonode auth api when `LOCKDOWN_GEONODE` is set to True.

load the docker_oauth_apps fixture during the execution of the common.sh

added some sed statements to allow for an automated deployment of geoserver with a oauth config to exchange docker setup.

load the default oauth_apps fixture from geonode. The vagrant setup doesnt need to have any sed statements to adjust files in the geoserver data directory, becuase the default localhost will work.

migration that occured when running makemigrations for the addition of the themes app.