Use boundless-vendor-libs rpm and remove upstream requirements.txt
The boundless-vendor-libs has been added to docker and vagrant configuration. The location of the libs is now in `/opt/boundless/vendor/lib`, the files have also been adjusted to add the postgresql installed at `/opt/boundless/vendor/libs`. The inclusion of the vendor-libs allowed for epel to be removed as a dependency. A settings variable, `GEOS_LIBRARY_PATH`, was added for docker config, since a custom location was used for geos.

There was a change from having the Dockerfile for exchange install all requirements and then have common.sh upgrade the dependencies by using the upstream requirements. A change recently to upstream was causing issues when installing the dependency for six and paver. The change that has been applied to docker setup is to remove the geonode entry in the exchange requirements.txt and then install the local repo at `/mnt/geonode` in common.sh. This also aligns the versions that are installed in our rpms and in the vagrant dev setup.

Adjust travis ci to use the new vendor-libs rpm and new geoserver.war
Also remove packages that are needed e.g. elasticsearch, postgresql and rabbitmq... along with epel and postgresl repos.