Changed "ng-if" to "ng-show" with "ng-hide".
The primary cause of this bug was that the message was showing
before Angular had the chance to parse the document.

Adding "ng-hide" to the div's class makes it invisible to the user
while the page was loading, and then "ng-show" is used to remove
the "ng-hide" class when applicable. But as that is done by the
Angular parser, the content is ready when the div is shown.