NODE-700 Manage and Edit style options are missing from Edit Data
The buttons are hidden when the preview viewer is react, the buttons were set to geoxt viewer only (geoext viewer is the old ol2 viewer).  An issue has been created in geonode-client [https://github.com/GeoNode/geonode-client/issues/145]. Modified the template to at least expose the option to manage the available styles, until the feature can be added to the ol3 react viewer.