Added Support for Local Registry Development.
The previous version of the registry docker setup
used the registry RPM to setup the registry environment. This version
of the docker file properly sets up the environment to use a clone
of the repository.  If the developer wishes then they can setup
a "local mount" of the repository and develop in a similar manner
to how exchange, maploom, and geonode are currently configured.