initial implementation of audit app in exchange
I have it logging login events to a django model and accessible for view only in django admin. The admin form does not have the ability to adjust, add or delete the contents. There is also an option to declare where to output the contents to a separate logfile (json). see default.py settings.

Currently this only logs login/logout events, but can be extended for CRUD operations also.