Ensured exchange static files take priority.
Add context and settings variables to handle GeoQuery search

Update site_base.html to correctly size GEOINT Logo

With the latest change to the Exchange core style
the GeoInt logo becomes distorted and doesn't fit
properly on the nav bar like it should. This change
is meant to restyle the logo so it sits correctly.