Change "Get WMS" link to use layer name not title
URL is using the title of the layer in the link,
but it should be using the GeoNode name of the
resource in GeoServer.

For Example, A layer with the Title: "Land Use",
and the Name: "landuse_30603f9f" appeared as:

https://exchange.boundlessgeo.io/geoserver/
geonode/Land%20Use/wms?REQUEST=GetCapabilities&

which is incorrect. Instead it should be:

https://exchange.boundlessgeo.io/geoserver/
geonode/landuse_30603f9f/ows?service=wms&
version=1.3.0&request=GetCapabilities