Fix workspace items sometimes displaying improperly
Previously the length of the display for the name of
the item in the workspace was explicitly defined by
a number of characters. Because that section doesn't
use a monospace font, Layer names that had a collection
of wider characters like 'W' would take up more space
than Layers with names that had smaller letters like
'l'.

This change forces an explicit size for the display,
regardless of the size of the characters, causing
the names to trim at the proper length every time.

Additionally, you can now hover over a layer name
to see the full name, so if it is cut off and you
need to differentiate between two layers, you can.